update insert operation in indel table class, store pos_coverage as integer not float

+++ anvio/tables/__init__.py
@@ -167,7 +167,7 @@ variable_nts_table_types             = [   'text'  ,    'text'   , 'numeric',   
 
 indels_table_name                    = 'indels'
 indels_table_structure               = ['sample_id', 'split_name', 'type', 'sequence', 'start_in_contig', 'start_in_split', 'length' , 'coverage', 'pos_coverage']
-indels_table_types                   = ['text'     , 'text'      , 'text', 'text'    , 'numeric'        , 'numeric'       , 'numeric', 'numeric' , 'numeric']
+indels_table_types                   = ['text'     , 'text'      , 'text', 'text'    , 'numeric'        , 'numeric'       , 'numeric', 'numeric' , 'integer']
 
 views_table_name                     = 'views'
 views_table_structure                = ['view_id', 'target_table']
